refactor: split Controller & Finder method in repository (SRP)
