+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/CarController.java
@@ -1,0 +1,58 @@
+package id.ac.ui.cs.advprog.eshop.controller;
+
+import id.ac.ui.cs.advprog.eshop.model.Car;
+import id.ac.ui.cs.advprog.eshop.service.CarServiceImpl;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@Controller
+@RequestMapping("/car")
+class CarController extends ProductController{
+    @Autowired
+    private CarServiceImpl carService;
+
+    @GetMapping("/createCar")
+    public String createCarPage(Model model){
+        Car car = new Car();
+        model.addAttribute("car", car);
+        return "CreateCar";
+    }
+
+    @PostMapping("/createCar")
+    public String createCarPost(@ModelAttribute Car car, Model model){
+        carService.create(car);
+        return "redirect:listCar";
+    }
+
+    @GetMapping("/listCar")
+    public String carListPage(Model model){
+        List<Car> cars = carService.findAll();
+        model.addAttribute("cars", cars);
+        return "CarList";
+    }
+
+    @GetMapping("/editCar/{carId}")
+    public String editCarPage(@PathVariable String carId, Model model){
+        Car car = carService.findById(carId);
+        model.addAttribute("car", car);
+        return "EditCar";
+    }
+
+    @PostMapping("/editCar")
+    public String editCarPost(@ModelAttribute Car car, Model model){
+        System.out.println(car.getCarId());
+        carService.update(car.getCarId(), car);
+
+        return "redirect:listCar";
+    }
+
+    @PostMapping("/deleteCar")
+    public String deleteCar(@RequestParam("carId") String carId){
+        carService.deleteCarById(carId);
+        return "redirect:listCar";
+    }
+}
