+++ src/main/java/id/ac/ui/cs/advprog/eshop/controller/ProductController.java
@@ -58,51 +58,3 @@ public class ProductController {
         return "redirect:/product/list";
     }
 }
-
-@Controller
-@RequestMapping("/car")
-class CarController extends ProductController{
-    @Autowired
-    private CarServiceImpl carService;
-
-    @GetMapping("/createCar")
-    public String createCarPage(Model model){
-        Car car = new Car();
-        model.addAttribute("car", car);
-        return "CreateCar";
-    }
-
-    @PostMapping("/createCar")
-    public String createCarPost(@ModelAttribute Car car, Model model){
-        carService.create(car);
-        return "redirect:listCar";
-    }
-
-    @GetMapping("/listCar")
-    public String carListPage(Model model){
-        List<Car> cars = carService.findAll();
-        model.addAttribute("cars", cars);
-        return "CarList";
-    }
-
-    @GetMapping("/editCar/{carId}")
-    public String editCarPage(@PathVariable String carId, Model model){
-        Car car = carService.findById(carId);
-        model.addAttribute("car", car);
-        return "EditCar";
-    }
-
-    @PostMapping("/editCar")
-    public String editCarPost(@ModelAttribute Car car, Model model){
-        System.out.println(car.getCarId());
-        carService.update(car.getCarId(), car);
-
-        return "redirect:listCar";
-    }
-
-    @PostMapping("/deleteCar")
-    public String deleteCar(@RequestParam("carId") String carId){
-        carService.deleteCarById(carId);
-        return "redirect:listCar";
-    }
-}
