+++ src/main/java/id/ac/ui/cs/advprog/eshop/repository/ProductRepository.java
@@ -2,6 +2,7 @@ package id.ac.ui.cs.advprog.eshop.repository;
 
 import id.ac.ui.cs.advprog.eshop.model.Product;
 import org.springframework.stereotype.Repository;
+import id.ac.ui.cs.advprog.eshop.repository.ProductFinder;
 
 import java.util.ArrayList;
 import java.util.Iterator;
@@ -10,6 +11,7 @@ import java.util.List;
 @Repository
 public class ProductRepository{
     private List<Product> productData = new ArrayList<>();
+    private ProductFinder productFinder = new ProductFinder();
 
     public Product create(Product product){
         if(product.getProductQuantity() < 0){
@@ -22,15 +24,6 @@ public class ProductRepository{
     public Product delete(Product product){
         productData.remove(product);
         return product;
-    }
-
-    public Product findById(String id){
-        for(Product product : productData){
-            if(product.getProductId().equals(id)){
-                return product;
-            }
-        }
-        return null;
     }
 
     public Product edit(Product updatedProduct){
@@ -47,7 +40,11 @@ public class ProductRepository{
         return null;
     }
 
+    public Product findById(String id){
+        return productFinder.findById(id, productData);
+    }
+
     public Iterator<Product> findAll(){
-        return productData.iterator();
+        return productFinder.findAll(productData);
     }
 }
